golang/18. 4Sum_test.go
package golang

import (
	"testing"

	"github.com/google/go-cmp/cmp"
)

func Test_fourSum(t *testing.T) {
	type args struct {
		nums   []int
		target int
	}
	tests := []struct {
		name string
		args args
		want [][]int
	}{
		{
			name: "e1",
			args: args{
				nums:   []int{1, 0, -1, 0, -2, 2},
				target: 0,
			},
			want: [][]int{{-1, 0, 0, 1}, {-2, -1, 1, 2}, {-2, 0, 0, 2}},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := fourSum(tt.args.nums, tt.args.target); !cmp.Equal(got, tt.want, sort2DIntSlice) {
				t.Errorf("fourSum() = %v, want %v", got, tt.want)
			}
		})
	}
}
